Add files via upload
Input the user’s first, middle, and last name as three separate strings and
display the name in the order of the first name, the middle initial, and the last
name. Include the period after the middle initial. If the input strings are
Wolfgang, Amadeus, and Mozart, for example, then the output would be
Wolfgang A. Mozart. Use the console window for output.